Order/requirements.txt
aniso8601==9.0.1
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
click==7.1.2
distlib==0.3.1
Django==3.1.5
djangorestframework==3.12.2
filelock==3.0.12
Flask==1.1.2
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.5.1
greenlet==1.0.0
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.3
kafka-python==2.0.2
lazy-object-proxy==1.4.3
mysql-connector==2.2.9
mysqlclient==2.0.3
mariadb==1.0.6
MarkupSafe==1.1.1
mccabe==0.6.1
pylint==2.6.0
pylint-flask==0.6
pylint-flask-sqlalchemy==0.2.0
pylint-plugin-utils==0.6
pytz==2020.5
six==1.15.0
SQLAlchemy==1.4.5
sqlparse==0.4.1
toml==0.10.2
virtualenv==20.4.3
Werkzeug==1.0.1
wrapt==1.12.1
